requirement.txt
os
pypdf2
pandas
langchain-community

